+++ src/components/calcite-icon/calcite-icon.tsx
@@ -13,6 +13,7 @@ import { getElementDir } from "../../utils/dom";
 import { fetchIcon, scaleToPx } from "./utils";
 import { Scale } from "../../interfaces/Icon";
 import { Theme } from "../../interfaces/common";
+import { CalciteIconPath, CalciteMultiPathEntry } from "@esri/calcite-ui-icons";
 
 @Component({
   assetsDirs: ["assets"],
@@ -35,14 +36,6 @@ export class CalciteIcon {
   //  Properties
   //
   //--------------------------------------------------------------------------
-
-  /**
-   * When true, the icon will be filled.
-   */
-  @Prop({
-    reflect: true,
-  })
-  filled: boolean = false;
 
   /**
    * The name of the icon to display. The value of this property must match the icon name from https://esri.github.io/calcite-ui-icons/.
@@ -113,7 +106,7 @@ export class CalciteIcon {
     const dir = getElementDir(el);
     const size = scaleToPx[scale];
     const semantic = !!textLabel;
-
+    const paths = [].concat(pathData || "");
     return (
       <Host
         aria-label={semantic ? textLabel : null}
@@ -130,7 +123,13 @@ export class CalciteIcon {
           width={size}
           viewBox={`0 0 ${size} ${size}`}
         >
-          <path d={pathData} />
+          {paths.map((path: string | CalciteMultiPathEntry) =>
+            typeof path === "string" ? (
+              <path d={path} />
+            ) : (
+              <path d={path.d} opacity={"opacity" in path ? path.opacity : 1} />
+            )
+          )}
         </svg>
       </Host>
     );
@@ -145,7 +144,7 @@ export class CalciteIcon {
   private intersectionObserver: IntersectionObserver;
 
   @State()
-  private pathData: string;
+  private pathData: CalciteIconPath;
 
   @State()
   private visible = false;
@@ -157,16 +156,15 @@ export class CalciteIcon {
   //--------------------------------------------------------------------------
 
   @Watch("icon")
-  @Watch("filled")
   @Watch("scale")
   private async loadIconPathData(): Promise<void> {
-    const { filled, icon, scale, visible } = this;
+    const { icon, scale, visible } = this;
 
     if (!Build.isBrowser || !icon || !visible) {
       return;
     }
 
-    this.pathData = await fetchIcon({ icon, scale, filled });
+    this.pathData = await fetchIcon({ icon, scale });
   }
 
   private waitUntilVisible(callback: () => void): void {